refactor: challenges string to strings
