+++ lib/gem_ruby.rb
@@ -14,8 +14,8 @@ require_relative "gem_ruby/algorithms/greedy/maximize_array.rb"
 require_relative "gem_ruby/algorithms/greedy/minimum_product.rb"
 require_relative "gem_ruby/algorithms/greedy/minimum_sum_product.rb"
 
-require_relative "gem_ruby/challenges/string/missing_characters.rb"
-require_relative "gem_ruby/challenges/string/count.rb"
+require_relative "gem_ruby/challenges/strings/missing_characters.rb"
+require_relative "gem_ruby/challenges/strings/count.rb"
 
 require_relative "gem_ruby/challenges/arrays/find_three.rb"
 require_relative "gem_ruby/challenges/arrays/reversal.rb"
