+++ lib/gem_ruby/challenges/strings/common_characters.rb
