+++ lib/gem_ruby/challenges/strings/group_words.rb
